third_party/WebKit/LayoutTests/fast/dom/shadow/cascade-of-treeboundary-crossing-rules-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test for casacde of treeboundary crossing rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS borderColorOf(getNodeInComposedTree("target")) is "rgb(0, 128, 0)"
PASS borderColorOf(getNodeInComposedTree("host/target")) is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
